requirements.txt
numpy
scipy
pandas
matplotlib
jupyter
ipython
scikit-learn
gensim
nltk

# local package
-e .

# external requirements
dvc[s3]
# jq
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

